fix: change request body of link patch api
Get updated data by array to update at once.
